Add CIDR notation flag to 'go-ip' cmd tool
In 'go-ip' command line tool, added the optional "cidr" flag to allow the user to input network address in CIDR notation. If the flag is present, the application will handle network information according to the input. This improves the user experience by providing more versatility in input methods, as well as a more informative output.