Add go.sum file with dependency versions
A go.sum file was added that includes versions for the application dependencies. This is intended to enhance reproducibility in building and launching the application, as well as improve dependency tracking. The file includes version information for numerous packages including multiple versions of cloud.google.com/go and github.com packages.